% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocks.r
\name{adapt.hf.blocks}
\alias{adapt.hf.blocks}
\title{Adapt header and footer of if blocks for output format
and parse already the if condition for faster runtime evaluation}
\usage{
adapt.hf.blocks(txt, block.df = NULL, out.type = "html",
  only.types = c("if", "note"), ...)
}
\description{
Adapt header and footer of if blocks for output format
and parse already the if condition for faster runtime evaluation
}

