% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunks.r
\name{args.to.chunk.header}
\alias{args.to.chunk.header}
\title{Create a Rmd chunk header line from a list of arguments}
\usage{
args.to.chunk.header(args = list(), type = "r", label = args$label)
}
\description{
Create a Rmd chunk header line from a list of arguments
}

