% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BibEntry_to_UTF8.R
\name{BibEntry_to_UTF8}
\alias{BibEntry_to_UTF8}
\title{Encode author names in UTF-8.}
\usage{
BibEntry_to_UTF8(refs)
}
\arguments{
\item{refs}{A BibEntry object.}
}
\value{
A BibEntry object.
}
\description{
Encode author names in BibEntry objects as UTF-8. Specially useful when working in Windows systems that do not support UTF-8.
}
\examples{
\dontrun{
library(knitcitations)
cleanbib()
cite_options(citation_format = "pandoc")
#citet("10.1111/nph.12929") # doesn't work
citep("10.1016/j.tree.2006.09.010")
citet("10.1111/j.1461-0248.2007.01060.x")
ref <- knitcitations:::get_bib()
ref.utf8 <- BibEntry_to_UTF8(ref)

}
}

