% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/capitalize_refs.R
\name{capitalize_authors}
\alias{capitalize_authors}
\title{Write Author Names in Title Case.}
\usage{
capitalize_authors(refs)
}
\arguments{
\item{refs}{A \code{\link[RefManageR]{BibEntry}} object.}
}
\value{
A BibEntry object with author names in Title Case.
}
\description{
This function capitalizes author names (given and family names), while setting all other characters to lower case.
}
\author{
F. Rodriguez-Sanchez
}
\seealso{
\code{\link{capitalize_refs}}
}

