% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/capitalize_refs.R
\name{capitalize_refs}
\alias{capitalize_refs}
\title{Fix capitalization of bibliographic references.}
\usage{
capitalize_refs(refs, TitleCase = FALSE)
}
\arguments{
\item{refs}{A \code{\link[RefManageR]{BibEntry}} object.}

\item{TitleCase}{Logical. Use Title Case (first letter of each word is capitalized)? Default is FALSE (only first letter of the title capitalized).}
}
\value{
A BibEntry object.
}
\description{
This function is just a wrapper of \code{\link{capitalize_authors}} and  \code{\link{capitalize_titles}}, aimed at fixing the capitalizacion of all references contained in a \code{\link[RefManageR]{BibEntry}} object all at once. Author names will appear with capitalized first letters, and titles will appear as Sentence case or Title Case, depending on \code{TitleCase} parameter.
}

