% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/capitalize_refs.R
\name{capitalize_titles}
\alias{capitalize_titles}
\title{Fix capitalization of reference titles.}
\usage{
capitalize_titles(refs, titlecase = FALSE)
}
\arguments{
\item{refs}{A \code{\link[RefManageR]{BibEntry}} object.}

\item{titlecase}{Logical. Use Title Case (first letter of each word is capitalized)? Default is FALSE (only first letter of the title capitalized).}
}
\value{
A BibEntry object with titles capitalized according to \code{titlecase}.
}
\description{
This function changes capitalization of reference titles in a \code{\link[RefManageR]{BibEntry}} object to Title Case (first letter of each word is capitalized), or only first word capitalized.
}
\author{
F. Rodriguez-Sanchez
}
\seealso{
\code{\link{capitalize_refs}}, and \code{\link[tools]{toTitleCase}} and \code{\link[Hmisc]{capitalize}} for details.
}

