% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd.r
\name{cat.rmd.params}
\alias{cat.rmd.params}
\title{Scan all used block and whisker parameters in an .rmd file
and create a template for a list}
\usage{
cat.rmd.params(file = NULL, text = readLines(file, warn = FALSE),
  use.blocks = TRUE, use.whiskers = TRUE)
}
\description{
Scan all used block and whisker parameters in an .rmd file
and create a template for a list
}

