% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunks.r
\name{chunk.opt.string.to.list}
\alias{chunk.opt.string.to.list}
\title{Translates a chunk header to a list of its option}
\usage{
chunk.opt.string.to.list(str, keep.name = FALSE)
}
\arguments{
\item{str}{the chunk header as written in the rmd file}

\item{keep.name}{shall the chunk name be kept?}
}
\description{
Translates a chunk header to a list of its option
}

