% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/note.r
\name{compile.note.block.to.tag}
\alias{compile.note.block.to.tag}
\title{extract #< if blocks from a rmd txt}
\usage{
compile.note.block.to.tag(txt, env = parent.frame(), info = NULL)
}
\description{
extract #< if blocks from a rmd txt
}

