% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocks.r
\name{compile.note.blocks.to.html}
\alias{compile.note.blocks.to.html}
\title{extract #< if blocks from a rmd txt}
\usage{
compile.note.blocks.to.html(txt, env = parent.frame(), call.list = NULL,
  block.df = NULL, del.rows.na = FALSE)
}
\description{
extract #< if blocks from a rmd txt
}

