% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocks.r
\name{compile.rmd.blocks}
\alias{compile.rmd.blocks}
\title{compile all given types of rmd blocks}
\usage{
compile.rmd.blocks(txt, to = "html", env = parent.frame(),
  call.list = NULL, block.df = NULL, types = c("if", "note"),
  use.del.rows.na = FALSE, replace.funs = NULL)
}
\description{
compile all given types of rmd blocks
}

