% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd.r
\name{compile.rmd}
\alias{compile.rmd}
\title{Main function to compile rmd to html}
\usage{
compile.rmd(file = NULL, text = readLines(file, warn = FALSE),
  params = NULL, parent.env = parent.frame(), use.blocks = FALSE,
  use.whiskers = TRUE, start.line = NULL, set.utf8 = TRUE,
  knit = !chunk.like.whisker, chunk.like.whisker = FALSE,
  out.type = "html", markdown.blocks.call.list = NULL,
  whiskers.call.list = NULL, fragment.only = TRUE)
}
\description{
Main function to compile rmd to html
}

