% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunks.r
\name{eval.chunks.in.text}
\alias{eval.chunks.in.text}
\title{Render all knitr chunks in the same way as a whisker
 (taking into some chunk options, like results="asis")}
\usage{
eval.chunks.in.text(rmd, params = list(), env = parent.frame())
}
\description{
Render all knitr chunks in the same way as a whisker
 (taking into some chunk options, like results="asis")
}

