% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ph.r
\name{eval.placeholder}
\alias{eval.placeholder}
\title{Evaluate a placeholder and return its value}
\usage{
eval.placeholder(ph, envir = parent.frame(), chunks = "knit",
  dir = getwd(), out.type = "html", cr = NULL, on.error = c("null",
  "error", "stop")[1], ...)
}
\description{
Evaluate a placeholder and return its value
}

