% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocks.r
\name{eval.rmd.blocks.in.text}
\alias{eval.rmd.blocks.in.text}
\title{compile all given types of rmd blocks}
\usage{
eval.rmd.blocks.in.text(txt, to = "html", env = parent.frame(),
  call.list = NULL, block.df = NULL, only.types = NULL,
  ignore.types = "if", use.del.rows.na = FALSE, replace.funs = NULL)
}
\description{
compile all given types of rmd blocks
}

