% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whisker.r
\name{eval.whiskers.in.text}
\alias{eval.whiskers.in.text}
\title{evaluate whiskers and replace them in the text}
\usage{
eval.whiskers.in.text(str, envir = parent.frame(),
  signif = getOption("whiskerSignifDigits"),
  round = getOption("whiskerRoundDigits"), add.params = TRUE,
  whiskers.call.list = NULL)
}
\description{
evaluate whiskers and replace them in the text
}

