% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nested.r
\name{find.dot.blocks}
\alias{find.dot.blocks}
\title{Find blocks that have only a starting line of the form}
\usage{
find.dot.blocks(txt, dot.levels = NULL, dot.start = "#. ")
}
\description{
#. type arguments
}

