% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown_utils.r
\name{find.rmd.chunks}
\alias{find.rmd.chunks}
\title{Find the start and end lines of all rmd chunks}
\usage{
find.rmd.chunks(rmd)
}
\arguments{
\item{rmd}{the rmd code as character vector, one element per line}
}
\description{
Find the start and end lines of all rmd chunks
}

