% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nested.r
\name{find.rmd.nested}
\alias{find.rmd.nested}
\title{find blocks, chunks and dot blocks and add nesting info}
\usage{
find.rmd.nested(txt, dot.levels = NULL)
}
\arguments{
\item{txt}{the rmd source as character vector, each line one element}

\item{dot.levels}{a list that describes the level of dot block types}
}
\value{
a data.frame
}
\description{
find blocks, chunks and dot blocks and add nesting info
}

