% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.r
\name{get.levels.parents.by.types}
\alias{get.levels.parents.by.types}
\title{Find the parent index of each row given a vector levels that describes the nestedness}
\usage{
get.levels.parents.by.types(levels, types,
  parent.types = setdiff(unique(types), c(NA)))
}
\arguments{
\item{levels}{integer vector of levels must start with lowest level and increase by 1 or
decrease to an integer number}

\item{types}{character vector of types}

\item{parent.types}{the parent.types that shall be characterized}
}
\value{
a matrix of parent type indices with length(levels) rows and length(parent.types) columns. If there is no parent type, we enter a 0.
}
\description{
Find the parent index of each row given a vector levels that describes the nestedness
}

