% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.r
\name{get.start.end.levels}
\alias{get.start.end.levels}
\title{Find the levels given ordered start and end positions of possible nested blocks}
\usage{
get.start.end.levels(start, end, start.level = 1L)
}
\arguments{
\item{start}{vector of start positions of the blocks}

\item{end}{vector of end positions of the blocks}

\item{start.level}{the initial level}
}
\value{
a vector of levels of each blocks, moore deeply nested blocks have higher levels
}
\description{
Find the levels given ordered start and end positions of possible nested blocks
}

