% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.r
\name{get.types.inds}
\alias{get.types.inds}
\title{Gets for each element the index of its type}
\usage{
get.types.inds(types)
}
\arguments{
\item{types}{a character vector of types}
}
\description{
Gets for each element the index of its type
}

