% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr.r
\name{hideShowButton}
\alias{hideShowButton}
\title{A button that toogles whether content in a div
is displayed or not}
\usage{
hideShowButton(id, label, content = NULL, div.id = NULL, shown = FALSE,
  ...)
}
\description{
A button that toogles whether content in a div
is displayed or not
}

