% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.r
\name{html.table}
\alias{html.table}
\title{own function to print a compact html table from a data frame
with option to select row}
\usage{
html.table(df, sel.row = NULL, col.names = TRUE, row.names = FALSE,
  border = TRUE, bg.color = c("#dddddd", "#ffffff"), font.size = "80\%",
  round.digits = 8, signif.digits = 8, col.tooltips = NULL, ...)
}
\description{
own function to print a compact html table from a data frame
with option to select row
}

