% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html.r
\name{html.template}
\alias{html.template}
\title{Variant of htmltools::htmlTemplate}
\usage{
html.template(file = NULL, html = NULL, envir = parent.frame(),
  document_ = "auto")
}
\description{
Variant of htmltools::htmlTemplate
}

