% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml.r
\name{import.yaml}
\alias{import.yaml}
\title{A more convenient yaml importer}
\usage{
import.yaml(file = NULL, text = NULL, verbose = FALSE,
  keep.quotes = FALSE, quote.char = "__QUOTE__", catch.error = TRUE,
  check.by.row = FALSE, utf8 = TRUE)
}
\description{
A more convenient yaml importer
}

