% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.r
\name{inline.dependencies}
\alias{inline.dependencies}
\title{Inline dependencies with local file references}
\usage{
inline.dependencies(deps, mustWork = TRUE)
}
\description{
Inline dependencies with local file references
}

