% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.r
\name{inline.dependency}
\alias{inline.dependency}
\title{Inline a dependency with local file references}
\usage{
inline.dependency(dependency, mustWork = TRUE)
}
\description{
Inline a dependency with local file references
}

