% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr.r
\name{knit.chunk}
\alias{knit.chunk}
\title{Knits the rmd txt inside a temporary directory instead of the current wd}
\usage{
knit.chunk(text, envir = parent.frame(), fragment.only = TRUE,
  quiet = TRUE, encoding = getOption("encoding"), html.table = TRUE,
  out.type = "html", knit.dir = getwd(), use.commonmark = TRUE,
  deps.action = c("add", "ignore")[1], args = NULL, eval_mode = c("knit",
  "sculpt", "eval")[1], show_code = c("no", "note", "open_note", "note_after",
  "open_note_after", "before", "after")[1], code.highlight = use.commonmark,
  ...)
}
\description{
Does not create /figure subfolder in current wd
}

