% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr.r
\name{knit.rmd.in.temp}
\alias{knit.rmd.in.temp}
\title{Knits the rmd txt inside a temporary directory instead of the current wd}
\usage{
knit.rmd.in.temp(text, envir = parent.frame(), ...)
}
\description{
Does not create /figure subfolder in current wd
}

