% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr.r
\name{knit.rmd}
\alias{knit.rmd}
\title{Knits the rmd txt}
\usage{
knit.rmd(text, envir = parent.frame(), fragment.only = TRUE, quiet = TRUE,
  encoding = getOption("encoding"), html.table = TRUE, out.type = "html",
  use.commonmark = FALSE)
}
\description{
Knits the rmd txt
}

