% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html.r
\name{md2html}
\alias{md2html}
\title{Own markdown to html converter that interfaces commonmark}
\usage{
md2html(text, fragment.only = TRUE, smart = FALSE, ...)
}
\arguments{
\item{text}{the variable containing the markdown text}

\item{fragment.only}{only a fragment or should html headers be added}

\item{smart}{smart punctuation (default FALSE)}
}
\value{
The created HTML as a string
}
\description{
Own markdown to html converter that interfaces commonmark
}

