% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.r
\name{normalize.id}
\alias{normalize.id}
\title{Normalize an id to letters that are allowed}
\usage{
normalize.id(str, allowed = c(letters, LETTERS, 0:9, "_"), subst = "_")
}
\description{
Normalize an id to letters that are allowed
}

