% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown_utils.r
\name{parse.chunk.names}
\alias{parse.chunk.names}
\title{Take a vector of chunk header lines and returns the chunk names}
\usage{
parse.chunk.names(header)
}
\arguments{
\item{header}{the chunk headers}
}
\description{
Take a vector of chunk header lines and returns the chunk names
}

