% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whisker.r
\name{paste.whiskers}
\alias{paste.whiskers}
\title{Subtitutes whiskers and pastes string parts together}
\usage{
paste.whiskers(str, values = parent.frame(), eval = TRUE,
  signif.digits = NULL, error.val = "`Error`", empty.val = NULL,
  whisker.start = "{{", whisker.end = "}}", sep = "", collapse = NULL,
  return.list = FALSE)
}
\description{
Unlike replace whiskers, returns a string vector if whiskers evaluate as vectors
Cannot deal with nested whiskers
}

