% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd.r
\name{remove.rmd.chunks}
\alias{remove.rmd.chunks}
\title{Removes the rmd chunks with the given chunk names from rmd code}
\usage{
remove.rmd.chunks(rmd, chunk.names)
}
\arguments{
\item{rmd}{the rmd code as character vector, one element per line}

\item{chunk.names}{the list of rmd chunks that shall be removed}
}
\description{
Removes the rmd chunks with the given chunk names from rmd code
}

