% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whisker.r
\name{render.chunk.like.whisker}
\alias{render.chunk.like.whisker}
\title{Render a knitr chunk in the same way as a whisker (taking into some chunk options, like results="asis")}
\usage{
render.chunk.like.whisker(code, call = NULL, options = NULL,
  env = parent.frame())
}
\description{
Render a knitr chunk in the same way as a whisker (taking into some chunk options, like results="asis")
}

