% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whisker.r
\name{render.chunks.like.whisker}
\alias{render.chunks.like.whisker}
\title{Render all knitr chunks in the same way as a whisker
 (taking into some chunk options, like results="asis")}
\usage{
render.chunks.like.whisker(rmd, params = list(), env = parent.frame())
}
\description{
Render all knitr chunks in the same way as a whisker
 (taking into some chunk options, like results="asis")
}

