% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd.r
\name{render.compiled.rmd}
\alias{render.compiled.rmd}
\title{Render a compiled rmd}
\usage{
render.compiled.rmd(cr = NULL, txt = cr$body, envir = parent.frame(),
  fragment.only = FALSE, chunks = c("knit", "eval")[1], out.type = if
  (is.null(cr$out.type)) "html" else cr$out.type, use.print = "none",
  overwrite.values = FALSE, on.error = "error", use.commonmark = TRUE)
}
\description{
Render a compiled rmd
}

