% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.r
\name{render.deps.as.singletons.tags}
\alias{render.deps.as.singletons.tags}
\title{Render a list of dependencies as a list of singleton head tags
TO DO: Some dependencies may
use local file names...
Those dependencies must be fully inlined...}
\usage{
render.deps.as.singletons.tags(deps, inline.local.files = TRUE)
}
\description{
Render a list of dependencies as a list of singleton head tags
TO DO: Some dependencies may
use local file names...
Those dependencies must be fully inlined...
}

