% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr.r
\name{render.rmd.in.temp}
\alias{render.rmd.in.temp}
\title{Render with RMarkdown::render the rmd txt inside a temporary directory instead of the current wd}
\usage{
render.rmd.in.temp(text, envir = parent.frame(), quiet = TRUE, ...)
}
\description{
Does not create /figure subfolder in current wd
}

