% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd.r
\name{render.value}
\alias{render.value}
\title{Render a value in a format specified by out.type}
\usage{
render.value(val, out.type = "html", ...)
}
\description{
Render a value in a format specified by out.type
}

