% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocks.r
\name{replace.if.blocks}
\alias{replace.if.blocks}
\title{extract #< if blocks from a rmd txt}
\usage{
replace.if.blocks(txt, env = parent.frame(), call.list = NULL,
  block.df = NULL, warn.if.na = TRUE, del.rows.na = FALSE)
}
\description{
extract #< if blocks from a rmd txt
}

