% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocks.r
\name{replace.rmd.blocks}
\alias{replace.rmd.blocks}
\title{replace all given types of rmd blocks}
\usage{
replace.rmd.blocks(txt, env = parent.frame(), call.list = NULL,
  block.df = NULL, types = c("note", "if"), use.del.rows.na = FALSE,
  replace.funs = NULL)
}
\description{
replace all given types of rmd blocks
}

