% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whisker.r
\name{replace.whiskers}
\alias{replace.whiskers}
\title{replace whiskers}
\usage{
replace.whiskers(str, env = parent.frame(),
  signif = getOption("whiskerSignifDigits"),
  round = getOption("whiskerRoundDigits"), add.params = TRUE,
  whiskers.call.list = NULL)
}
\description{
replace whiskers
}

