% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ph.r
\name{reverse.whisker.placeholders}
\alias{reverse.whisker.placeholders}
\title{Reverses whisker placeholders by their original whiskers}
\usage{
reverse.whisker.placeholders(txt, ph = cr$ph, cr = NULL)
}
\description{
May be useful if blocks evaluate whiskers themselves or
after rmd to html transformation
}

