% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd.r
\name{rmd.between.start.end.lines}
\alias{rmd.between.start.end.lines}
\title{Extract rmd txt between start.line and end.line tag}
\usage{
rmd.between.start.end.lines(txt, start.line = "<!-- START -->",
  end.line = "<!-- END -->", return.start.end = FALSE)
}
\description{
Extract rmd txt between start.line and end.line tag
}

