% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html.r
\name{rmd.blocks.to.placeholder}
\alias{rmd.blocks.to.placeholder}
\title{Replace blocks with placeholders of the form {{id}}}
\usage{
rmd.blocks.to.placeholder(txt, block.df = NULL, whisker.prefix = "{{",
  whisker.postfix = "}}", del.rows.na = FALSE)
}
\description{
Replace blocks with placeholders of the form {{id}}
}

