% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html.r
\name{rmd.blocks.to.placeholders}
\alias{rmd.blocks.to.placeholders}
\title{Replace blocks with placeholderss of the form {{id}}}
\usage{
rmd.blocks.to.placeholders(txt, block.df = NULL, whisker.prefix = "{{",
  whisker.postfix = "}}", del.rows.na = FALSE, ignore.types = NULL,
  only.types = NULL, add.info = TRUE, ...)
}
\description{
TODO: Need to deal with nested blocks: replace from inner to outer
}

