% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html.r
\name{rmd.chunks.to.placeholder}
\alias{rmd.chunks.to.placeholder}
\title{Replace chunks with placeholders of the form {{id}}}
\usage{
rmd.chunks.to.placeholder(txt, whisker.prefix = "{{",
  whisker.postfix = "}}", del.rows.na = FALSE)
}
\description{
Replace chunks with placeholders of the form {{id}}
}

