% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html.r
\name{rmd.whiskers.to.placeholder}
\alias{rmd.whiskers.to.placeholder}
\title{Replace whiskers with placeholders of the form {{id}}}
\usage{
rmd.whiskers.to.placeholder(txt)
}
\description{
Replace whiskers with placeholders of the form {{id}}
}

