% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ph.r
\name{rmd.whiskers.to.placeholders}
\alias{rmd.whiskers.to.placeholders}
\title{Replace whiskers with placeholderss of the form {{id}}}
\usage{
rmd.whiskers.to.placeholders(txt, whisker.prefix = "{{",
  whisker.postfix = "}}", add.info = TRUE)
}
\description{
Replace whiskers with placeholderss of the form {{id}}
}

