% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yamlrmd.r
\name{split.by.line.start}
\alias{split.by.line.start}
\title{Splits a text vector into different blocks by a start line token}
\usage{
\method{split}{by.line.start}(txt, start.with = NULL, block.lines = NULL,
  add.start = TRUE, merge.lines = TRUE)
}
\value{
a data frame
}
\description{
Splits a text vector into different blocks by a start line token
}

