% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_print.r
\name{table.knit_print.data.frame}
\alias{table.knit_print.data.frame}
\title{knitr data.frame printer as nice HTML table with several options}
\usage{
table.knit_print.data.frame(x, table.max.rows = 100, round.digits = 8,
  signif.digits = 8, html.data.frame = TRUE, show.col.tooltips = TRUE,
  col.tooltips = NULL, output = "html", options = NULL, ...)
}
\description{
knitr data.frame printer as nice HTML table with several options
}

