% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html.r
\name{view.html}
\alias{view.html}
\title{View an extended rmd file}
\usage{
view.html(file = NULL, text = if (!is.null(file)) readLines(file, warn =
  FALSE) else NULL, ui = NULL, browser = rstudio::viewer, dir = getwd())
}
\description{
View an extended rmd file
}

