% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd.r
\name{view.rmd}
\alias{view.rmd}
\title{View an extended rmd file}
\usage{
view.rmd(file = NULL, text = readLines(file, warn = FALSE),
  envir = list(), use.blocks = FALSE, use.whiskers = TRUE,
  start.line = "<!-- START -->", end.line = "<!-- END -->",
  set.utf8 = TRUE, knit = !chunks.like.whisker,
  chunks.like.whisker = FALSE, out.type = "shiny")
}
\description{
View an extended rmd file
}

