% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/write_bibtex_enhanced.R
\name{write_bibtex_enhanced}
\alias{write_bibtex_enhanced}
\title{Fix references and write BibTeX file.}
\usage{
write_bibtex_enhanced(entry = NULL, file = "references.bib",
  append = FALSE, utf8 = FALSE, capitalize = TRUE, ...)
}
\arguments{
\item{entry}{a 'bibentry' object or list of bibentry objects.
If NULL, writes all that have currently been cited.}

\item{file}{output bibtex file. Will automatically append '.bib' if not
added. if 'NULL' will use stdout.}

\item{append}{a logical indicating that bibtex entries be added the the
file.  If FALSE (default), the file is overwritten.}

\item{utf8}{Logical. If TRUE, convert author names to UTF-8 encoding.}

\item{capitalize}{Logical. If TRUE, fix capitalization of references (see \code{\link{capitalize_refs}}).}

\item{...}{additional arguments to WriteBib}
}
\value{
a list of citation information, invisibly
}
\description{
An enhanced version of \code{\link[knitcitations]{write.bibtex}} aimed at fixing capitalization styles of references, and overcoming problems with non-UTF-8 encoding (particularly in Windows systems).
}
\examples{
write_bibtex_utf8(c(citation("knitr"),
                citation("knitcitations"),
                citation("RCurl")))
}
\author{
Carl Boettiger, modified by F. Rodriguez-Sanchez.
}
\seealso{
\code{\link{capitalize_refs}} and \code{\link{BibEntry_to_UTF8}}.
}

