% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/write_bibtex_windows.R
\name{write_bibtex_windows}
\alias{write_bibtex_windows}
\title{Encode authors names in UTF-8 and write BibTeX file.}
\usage{
write_bibtex_windows(entry = NULL, file = "references.bib",
  append = FALSE, ...)
}
\arguments{
\item{entry}{a 'bibentry' object or list of bibentry objects.
If NULL, writes all that have currently been cited.}

\item{file}{output bibtex file. Will automatically append '.bib' if not
added. if 'NULL' will use stdout.}

\item{append}{a logical indicating that bibtex entries be added the the
file.  If FALSE (default), the file is overwritten.}

\item{...}{additional arguments to WriteBib}
}
\value{
a list of citation information, invisibly
}
\description{
Modified from \code{\link[knitcitations]{write.bibtex}} to overcome problems with non-UTF-8 encoding in Windows systems.
}
\examples{
write_bibtex_windows(c(citation("knitr"),
                citation("knitcitations"),
                citation("RCurl")))
}
\author{
Carl Boettiger, slightly modified by F. Rodriguez-Sanchez.
}

