% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/capitalize_authors.R
\name{capitalize_authors}
\alias{capitalize_authors}
\title{Write Author Names in Title Case.}
\usage{
capitalize_authors(ref)
}
\arguments{
\item{ref}{A \code{\link[RefManageR]{BibEntry}} object.}
}
\value{
A BibEntry object with author names in Title Case.
}
\description{
This function capitalizes author names (given and family names), while setting all other characters to lower case.
}
\author{
F. Rodriguez-Sanchez
}

