% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd.r
\name{compile.rmd}
\alias{compile.rmd}
\title{Main function to compile rmd to html}
\usage{
compile.rmd(file = NULL, text = readLines(file, warn = FALSE),
  envir = list(), if.blocks = c("ph", "render", "ignore")[1],
  blocks = c("ph", "render", "ignore")[1], whiskers = c("ph", "render",
  "render.as.text", "ignore")[1], chunks = c("ph", "knit", "render",
  "ignore")[1], start.line = "<!-- START -->", end.line = "<!-- END -->",
  set.utf8 = TRUE, out.type = "html", fragment.only = FALSE,
  whiskers.call.list = NULL, blocks.call.list = NULL, add.info = TRUE,
  use.commonmark = use.commonmark)
}
\description{
Main function to compile rmd to html
}

