% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whisker.r
\name{eval.whiskers.in.text}
\alias{eval.whiskers.in.text}
\title{replace whiskers}
\usage{
eval.whiskers.in.text(str, env = parent.frame(),
  signif = getOption("whiskerSignifDigits"),
  round = getOption("whiskerRoundDigits"), add.params = TRUE,
  whiskers.call.list = NULL)
}
\description{
replace whiskers
}

