% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocks.r
\name{find.rmd.blocks}
\alias{find.rmd.blocks}
\title{Find all rmd blocks that start with a line `#< ...` and end with a line `#>`}
\usage{
find.rmd.blocks(txt)
}
\arguments{
\item{txt}{the rmd code, separated into lines}
}
\description{
Find all rmd blocks that start with a line `#< ...` and end with a line `#>`
}

