% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr.r
\name{knit.chunk}
\alias{knit.chunk}
\title{Knits the rmd txt inside a temporary directory instead of the current wd}
\usage{
knit.chunk(text, envir = parent.frame(), fragment.only = TRUE,
  quiet = TRUE, encoding = getOption("encoding"), html.table = TRUE,
  out.type = "html", knit.dir = tempdir(), use.commonmark = TRUE)
}
\description{
Does not create /figure subfolder in current wd
}

