% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocks.r
\name{parse.block.args}
\alias{parse.block.args}
\title{Parse the name of an rmd block}
\usage{
parse.block.args(header, arg.str = NULL, add.type = TRUE, type = "")
}
\description{
Parse the name of an rmd block
}

