% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html.r
\name{rmd.chunks.to.placeholders}
\alias{rmd.chunks.to.placeholders}
\title{Replace chunks with placeholderss of the form {{id}}}
\usage{
rmd.chunks.to.placeholders(txt, whisker.prefix = "{{",
  whisker.postfix = "}}", del.rows.na = FALSE, add.info = TRUE)
}
\description{
Replace chunks with placeholderss of the form {{id}}
}

