% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd.r
\name{view.rmd}
\alias{view.rmd}
\title{View an extended rmd file}
\usage{
view.rmd(file = NULL, text = readLines(file, warn = FALSE), params = NULL,
  parent.env = parent.frame(), use.blocks = FALSE, use.whiskers = TRUE,
  start.line = NULL, set.utf8 = TRUE, knit = !chunk.like.whisker,
  chunk.like.whisker = FALSE)
}
\description{
View an extended rmd file
}

